fix reaction

+++ Lagrange.Core/Message/MessageChain.cs
@@ -26,9 +26,9 @@ public sealed class MessageChain : List<IMessageEntity>
 
     public MessageStyle? Style { get; internal set; }
 
-    public uint Sequence { get; } // for C2C message, it's the sequence of the message, for group message, it's the sequence of the group message
+    public ulong Sequence { get; } // for C2C message, it's the sequence of the message, for group message, it's the sequence of the group message
 
-    public uint ClientSequence { get; } // only for C2C message
+    public ulong ClientSequence { get; } // only for C2C message
 
     #region Internal Properties
 
@@ -46,14 +46,14 @@ public sealed class MessageChain : List<IMessageEntity>
 
     #endregion
 
-    internal MessageChain(uint friendUin, string selfUid, string friendUid, uint targetUin = 0, uint sequence = 0, uint clientSequence = 0, ulong? messageId = null,
+    internal MessageChain(uint friendUin, string selfUid, string friendUid, uint targetUin = 0, ulong sequence = 0, ulong clientSequence = 0, ulong? messageId = null,
         MessageType type = MessageType.Friend, uint appid = 0)
     {
         GroupUin = null;
         FriendUin = friendUin;
         TargetUin = targetUin;
         Sequence = sequence; // unuseful at there
-        ClientSequence = clientSequence == 0 ? (uint)Random.Shared.Next(100000000, int.MaxValue) : clientSequence;
+        ClientSequence = clientSequence == 0 ? (ulong)Random.Shared.Next(100000000, int.MaxValue) : clientSequence;
         SelfUid = selfUid;
         Uid = friendUid;
         MessageId = messageId ?? (0x01000000ul << 32) | (uint)Random.Shared.Next(100000000, int.MaxValue);
@@ -72,7 +72,7 @@ public sealed class MessageChain : List<IMessageEntity>
         Elements = new List<Elem>();
     }
 
-    internal MessageChain(uint groupUin, uint friendUin, uint sequence, ulong messageId = 0,uint appid = 0)
+    internal MessageChain(uint groupUin, uint friendUin, ulong sequence, ulong messageId = 0,uint appid = 0)
     {
         GroupUin = groupUin;
         FriendUin = friendUin;
